requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
openai
pypdf
python-docx